ESP32 initial port
all code ported to ESP32, feature complete and tested compatible with ESP8266. modifications made to library code,  examples, and travis build.

The ESP32 and '8266 frameworks are extremely close. the biggest incompatibility was the ip_addr structure (ip address) changes to provide for ipv6 by renaming/removing some v4 methods. The new code assumes ipv4 only for now, that will require work later.

The one thing that could be a lingering issue is the SDK changes to TCP server writer streams which removed the .writeAvailable() method. While it's nice to pretend the TX buffer is infinite in size, in extreme circumstances that will lead to dropped packets and memory crashes so I might need to dig underneath the Arduino API and write a more direct TCP transport to get that knowledge back.

A couple of small changes were made to the examples and library include header so that platform-specific includes can be avoided in sketch files.